Backend/requirements.txt
fastapi
uvicorn
python-multipart
torch
torchvision
Pillow
numpy<2
easyocr
